# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Amiri Phone - Algeria Ecommerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: featured
          in: query
          schema:
            type: boolean
        - name: onSale
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a single product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: string
        nameFr:
          type: string
        description:
          type: string
        descriptionAr:
          type: string
        descriptionFr:
          type: string
        price:
          type: number
        originalPrice:
          type: number
          nullable: true
        images:
          type: array
          items:
            type: string
        categoryId:
          type: integer
          nullable: true
        category:
          $ref: "#/components/schemas/Category"
          nullable: true
        stock:
          type: integer
        featured:
          type: boolean
        onSale:
          type: boolean
        discount:
          type: number
          nullable: true
        brand:
          type: string
          nullable: true
        model:
          type: string
          nullable: true
        specs:
          type: object
          nullable: true
          additionalProperties: true
        averageRating:
          type: number
          nullable: true
        reviewCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, price, images, stock, featured, onSale, reviewCount, createdAt, updatedAt]
    CreateProductRequest:
      type: object
      properties:
        name:
          type: string
        nameAr:
          type: string
        nameFr:
          type: string
        description:
          type: string
        descriptionAr:
          type: string
        descriptionFr:
          type: string
        price:
          type: number
        originalPrice:
          type: number
          nullable: true
        images:
          type: array
          items:
            type: string
        categoryId:
          type: integer
          nullable: true
        stock:
          type: integer
        featured:
          type: boolean
        onSale:
          type: boolean
        discount:
          type: number
          nullable: true
        brand:
          type: string
          nullable: true
        model:
          type: string
          nullable: true
        specs:
          type: object
          nullable: true
          additionalProperties: true
      required: [name, price, images, stock]
    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required: [products, total, limit, offset]
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: string
          nullable: true
        nameFr:
          type: string
          nullable: true
        slug:
          type: string
        icon:
          type: string
          nullable: true
        productCount:
          type: integer
      required: [id, name, slug, productCount]
...